# Searcher: fix optimization for empty patterns