In #59331, we refactored the searcher logic to support AND/ OR patterns in
searcher. While doing so, we accidentally regressed on an optimization that
avoids loading file content when the pattern is empty.

We now have a benchmark for empty patterns, but this benchmark was added
recently and wasn't run during that refactor.

Note: this will be backported to 5.3.

## Test plan

Ran new benchmarks (see below commment)