@jtibshirani 2024-02-20
Benchmark results

<details>

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/cmd/searcher/internal/search
                                                │   old.txt    │               new.txt               │
                                                │    sec/op    │    sec/op     vs base               │
SearchRegex_large_fixed-10                        5.380m ± ∞ ¹   5.365m ± ∞ ¹        ~ (p=0.151 n=5)
SearchRegex_rare_fixed-10                         7.138m ± ∞ ¹   7.045m ± ∞ ¹        ~ (p=0.310 n=5)
SearchRegex_large_fixed_casesensitive-10          2.864m ± ∞ ¹   3.082m ± ∞ ¹        ~ (p=0.310 n=5)
SearchRegex_large_empty_pattern-10                5.236m ± ∞ ¹   2.911m ± ∞ ¹  -44.40% (p=0.008 n=5)
SearchRegex_large_lang_filter_common-10           4.381m ± ∞ ¹   4.550m ± ∞ ¹   +3.87% (p=0.032 n=5)
SearchRegex_large_lang_filter_uncommon-10         1.215m ± ∞ ¹   1.206m ± ∞ ¹        ~ (p=0.548 n=5)
SearchRegex_large_re_dotstar-10                   224.9m ± ∞ ¹   225.3m ± ∞ ¹        ~ (p=1.000 n=5)
SearchRegex_large_re_common-10                    5.327m ± ∞ ¹   5.530m ± ∞ ¹   +3.82% (p=0.008 n=5)
SearchRegex_large_re_anchor-10                    56.81m ± ∞ ¹   57.99m ± ∞ ¹   +2.06% (p=0.008 n=5)
SearchRegex_large_capture_group-10                139.1m ± ∞ ¹   139.6m ± ∞ ¹        ~ (p=0.421 n=5)
SearchRegex_large_path/path_only-10               750.2µ ± ∞ ¹   738.7µ ± ∞ ¹        ~ (p=0.421 n=5)
SearchRegex_large_path/content_only-10            6.963m ± ∞ ¹   6.939m ± ∞ ¹        ~ (p=0.841 n=5)
SearchRegex_large_path/both_path_and_content-10   7.018m ± ∞ ¹   6.940m ± ∞ ¹        ~ (p=0.151 n=5)
SearchRegex_small_fixed-10                        142.1µ ± ∞ ¹   143.6µ ± ∞ ¹        ~ (p=0.095 n=5)
SearchRegex_small_fixed_casesensitive-10          98.90µ ± ∞ ¹   99.14µ ± ∞ ¹        ~ (p=0.095 n=5)
SearchRegex_small_re_dotstar-10                   2.258m ± ∞ ¹   2.246m ± ∞ ¹        ~ (p=0.151 n=5)
SearchRegex_small_re_common-10                    103.1µ ± ∞ ¹   102.8µ ± ∞ ¹        ~ (p=0.548 n=5)
SearchRegex_small_re_anchor-10                    725.3µ ± ∞ ¹   725.4µ ± ∞ ¹        ~ (p=1.000 n=5)
SearchRegex_small_capture_group-10                1.575m ± ∞ ¹   1.581m ± ∞ ¹        ~ (p=0.690 n=5)
geomean                                           3.201m         3.125m         -2.38%
¹ need >= 6 samples for confidence interval at level 0.95

                                                │    old.txt     │               new.txt                │
                                                │      B/op      │     B/op       vs base               │
SearchRegex_large_fixed-10                         14.97Mi ± ∞ ¹   14.97Mi ± ∞ ¹        ~ (p=0.548 n=5)
SearchRegex_rare_fixed-10                          14.97Mi ± ∞ ¹   14.97Mi ± ∞ ¹        ~ (p=0.690 n=5)
SearchRegex_large_fixed_casesensitive-10           14.83Ki ± ∞ ¹   14.93Ki ± ∞ ¹        ~ (p=0.516 n=5)
SearchRegex_large_empty_pattern-10                16.005Mi ± ∞ ¹   1.107Mi ± ∞ ¹  -93.08% (p=0.008 n=5)
SearchRegex_large_lang_filter_common-10            14.93Mi ± ∞ ¹   14.92Mi ± ∞ ¹        ~ (p=0.222 n=5)
SearchRegex_large_lang_filter_uncommon-10          14.97Mi ± ∞ ¹   14.95Mi ± ∞ ¹        ~ (p=0.095 n=5)
SearchRegex_large_re_dotstar-10                    599.9Mi ± ∞ ¹   599.8Mi ± ∞ ¹        ~ (p=0.151 n=5)
SearchRegex_large_re_common-10                     7.147Mi ± ∞ ¹   7.148Mi ± ∞ ¹        ~ (p=0.841 n=5)
SearchRegex_large_re_anchor-10                     4.691Mi ± ∞ ¹   4.700Mi ± ∞ ¹        ~ (p=0.310 n=5)
SearchRegex_large_capture_group-10                 3.290Mi ± ∞ ¹   3.277Mi ± ∞ ¹        ~ (p=0.690 n=5)
SearchRegex_large_path/path_only-10                8.577Ki ± ∞ ¹   8.572Ki ± ∞ ¹        ~ (p=0.357 n=5)
SearchRegex_large_path/content_only-10             64.67Ki ± ∞ ¹   65.07Ki ± ∞ ¹        ~ (p=0.738 n=5)
SearchRegex_large_path/both_path_and_content-10    63.12Ki ± ∞ ¹   63.12Ki ± ∞ ¹        ~ (p=0.690 n=5)
SearchRegex_small_fixed-10                         323.8Ki ± ∞ ¹   336.2Ki ± ∞ ¹   +3.83% (p=0.008 n=5)
SearchRegex_small_fixed_casesensitive-10           10.57Ki ± ∞ ¹   10.57Ki ± ∞ ¹        ~ (p=0.421 n=5)
SearchRegex_small_re_dotstar-10                    3.936Mi ± ∞ ¹   3.935Mi ± ∞ ¹   -0.03% (p=0.032 n=5)
SearchRegex_small_re_common-10                     48.43Ki ± ∞ ¹   48.38Ki ± ∞ ¹        ~ (p=0.151 n=5)
SearchRegex_small_re_anchor-10                     26.14Ki ± ∞ ¹   26.08Ki ± ∞ ¹   -0.23% (p=0.016 n=5)
SearchRegex_small_capture_group-10                 24.70Ki ± ∞ ¹   24.72Ki ± ∞ ¹        ~ (p=0.968 n=5)
geomean                                            800.8Ki         697.4Ki        -12.92%
¹ need >= 6 samples for confidence interval at level 0.95

                                                │   old.txt    │               new.txt                │
                                                │  allocs/op   │  allocs/op    vs base                │
SearchRegex_large_fixed-10                         260.0 ± ∞ ¹    260.0 ± ∞ ¹       ~ (p=0.595 n=5)
SearchRegex_rare_fixed-10                          677.0 ± ∞ ¹    677.0 ± ∞ ¹       ~ (p=1.000 n=5)
SearchRegex_large_fixed_casesensitive-10           211.0 ± ∞ ¹    211.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_large_empty_pattern-10                 122.0 ± ∞ ¹    110.0 ± ∞ ¹  -9.84% (p=0.008 n=5)
SearchRegex_large_lang_filter_common-10            318.0 ± ∞ ¹    319.0 ± ∞ ¹       ~ (p=0.762 n=5)
SearchRegex_large_lang_filter_uncommon-10          347.0 ± ∞ ¹    346.0 ± ∞ ¹       ~ (p=0.190 n=5)
SearchRegex_large_re_dotstar-10                   6.099M ± ∞ ¹   6.099M ± ∞ ¹       ~ (p=0.095 n=5)
SearchRegex_large_re_common-10                    68.79k ± ∞ ¹   68.79k ± ∞ ¹       ~ (p=0.111 n=5)
SearchRegex_large_re_anchor-10                    50.16k ± ∞ ¹   50.16k ± ∞ ¹       ~ (p=0.563 n=5)
SearchRegex_large_capture_group-10                25.28k ± ∞ ¹   25.27k ± ∞ ¹       ~ (p=0.444 n=5)
SearchRegex_large_path/path_only-10                129.0 ± ∞ ¹    129.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_large_path/content_only-10             669.0 ± ∞ ¹    669.0 ± ∞ ¹       ~ (p=0.444 n=5)
SearchRegex_large_path/both_path_and_content-10    647.0 ± ∞ ¹    647.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_fixed-10                         146.0 ± ∞ ¹    146.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_fixed_casesensitive-10           139.0 ± ∞ ¹    139.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_re_dotstar-10                   40.16k ± ∞ ¹   40.16k ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_re_common-10                     547.0 ± ∞ ¹    547.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_re_anchor-10                     351.0 ± ∞ ¹    351.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
SearchRegex_small_capture_group-10                 332.0 ± ∞ ¹    332.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                           1.419k         1.411k        -0.54%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```
</details>

Commands:
```
cd cmd/searcher/internal/search
go test -run '^$' -bench SearchRegex -count 5 > new.txt
git checkout d8f456c1c4446eb84b57c84832e199980d6a4245
go test -run '^$' -bench SearchRegex -count 5 > old.txt
benchstat old.txt new.txt > cmp.txt
```